Added in new variable
New variable reflecting the change in Theta from trial to trial. This will be used in place of p_rt as this will account for the changes in task difficulty from trial to trial. Currently, p_rt won't reflect this change.
All this has been added to and explained in 2_
Also, in 7a_ there is a new model to be run which includes this new variable as a fixed effect for now.